Split render-cuda.cpp into smaller files (#2334)
* render-cuda split, compile errors galore due to missing includes etc.

* render-cuda split and fully compiles

* Ran premake.bat to disable cuda; Added all new files

* Removed render-cuda files

* CI fixes

* Rerun CI
