+++ tools/gfx/cuda/cuda-pipeline-state.cpp
@@ -1,0 +1,22 @@
+// cuda-pipeline-state.cpp
+#include "cuda-pipeline-state.h"
+
+namespace gfx
+{
+#ifdef GFX_ENABLE_CUDA
+using namespace Slang;
+
+namespace cuda
+{
+
+void ComputePipelineStateImpl::init(const ComputePipelineStateDesc& inDesc)
+{
+    PipelineStateDesc pipelineDesc;
+    pipelineDesc.type = PipelineType::Compute;
+    pipelineDesc.compute = inDesc;
+    initializeBase(pipelineDesc);
+}
+
+} // namespace cuda
+#endif
+} // namespace gfx
