--- tools/gfx/cuda/render-cuda.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "../renderer-shared.h"
-
-namespace gfx
-{
-
-SlangResult SLANG_MCALL createCUDADevice(const IDevice::Desc* desc, IDevice** outDevice);
-}
