+++ tools/gfx/render.cpp
@@ -2,7 +2,6 @@
 #include "renderer-shared.h"
 #include "../../source/core/slang-math.h"
 #include "open-gl/render-gl.h"
-#include "cuda/render-cuda.h"
 #include "cpu/render-cpu.h"
 #include "debug-layer.h"
 
@@ -14,6 +13,7 @@ using namespace Slang;
 Result SLANG_MCALL createD3D11Device(const IDevice::Desc* desc, IDevice** outDevice);
 Result SLANG_MCALL createD3D12Device(const IDevice::Desc* desc, IDevice** outDevice);
 Result SLANG_MCALL createVKDevice(const IDevice::Desc* desc, IDevice** outDevice);
+Result SLANG_MCALL createCUDADevice(const IDevice::Desc* desc, IDevice** outDevice);
 
 static bool debugLayerEnabled = false;
